First bucket by attribute hash, then metric names

+++ newrelic-agent/src/main/java/com/newrelic/agent/stats/dimensional/Measure.java
@@ -1,7 +1,7 @@
 package com.newrelic.agent.stats.dimensional;
 
 interface Measure {
-    default void incrementCount(int count) {}
+    default void incrementCount(long count) {}
 
     default void addToSummary(double value) {}
 }
